+++ deps.ts
@@ -3,7 +3,5 @@ export {
   ResinCssEmitter,
   ResinCssGlobalStyle,
 } from 'https://deno.land/x/resin@v0.1.6/mod.ts';
-export { default as IconLoader2 } from 'https://deno.land/x/tabler_icons_tsx@0.0.4/tsx/loader-2.tsx';
-export { default as IconX } from 'https://deno.land/x/tabler_icons_tsx@0.0.4/tsx/x.tsx';
 export { DateTime, datetime } from 'https://deno.land/x/ptera@v1.0.2/mod.ts';
 export * as classNames from 'https://deno.land/x/classnames@0.1.1/index.ts';